refactor: reorganise internal `runtime/` directory
