+++ src/runtime/index.ts
@@ -1,5 +1,5 @@
 export { createImage } from './image'
-export { defineProvider } from './provider'
+export { defineProvider } from './utils/provider'
 export { createMapper, createOperationsGenerator } from './utils/index'
 
 export type { Mapper } from './utils/index'
