+++ src/runtime/plugins/image.ts
